+++ packages/partysocket/src/ws.ts
@@ -7,7 +7,60 @@
  * License MIT
  */
 
-import * as Events from "./events";
+export class Event {
+  public target: any;
+  public type: string;
+  constructor(type: string, target: any) {
+    this.target = target;
+    this.type = type;
+  }
+}
+
+export class ErrorEvent extends Event {
+  public message: string;
+  public error: Error;
+  constructor(error: Error, target: any) {
+    super("error", target);
+    this.message = error.message;
+    this.error = error;
+  }
+}
+
+export class CloseEvent extends Event {
+  public code: number;
+  public reason: string;
+  public wasClean = true;
+  constructor(code = 1000, reason = "", target: any) {
+    super("close", target);
+    this.code = code;
+    this.reason = reason;
+  }
+}
+export interface WebSocketEventMap {
+  close: CloseEvent;
+  error: ErrorEvent;
+  message: MessageEvent;
+  open: Event;
+}
+
+const Events = {
+  Event,
+  ErrorEvent,
+  CloseEvent,
+};
+
+export interface WebSocketEventListenerMap {
+  close: (
+    event: CloseEvent
+  ) => void | { handleEvent: (event: CloseEvent) => void };
+  error: (
+    event: ErrorEvent
+  ) => void | { handleEvent: (event: ErrorEvent) => void };
+  message: (
+    event: MessageEvent
+  ) => void | { handleEvent: (event: MessageEvent) => void };
+  open: (event: Event) => void | { handleEvent: (event: Event) => void };
+}
 
 function assert(condition: unknown, msg?: string): asserts condition {
   if (!condition) {
@@ -27,10 +80,6 @@ function isWebSocket(w: unknown): w is WebSocket {
     w.CLOSING === 2
   );
 }
-
-export type Event = Events.Event;
-export type ErrorEvent = Events.ErrorEvent;
-export type CloseEvent = Events.CloseEvent;
 
 export type Options = {
   // WebSocket?: any;
@@ -68,10 +117,10 @@ export type ProtocolsProvider =
 export type Message = string | ArrayBuffer | Blob | ArrayBufferView;
 
 export type ListenersMap = {
-  error: Events.WebSocketEventListenerMap["error"][];
-  message: Events.WebSocketEventListenerMap["message"][];
-  open: Events.WebSocketEventListenerMap["open"][];
-  close: Events.WebSocketEventListenerMap["close"][];
+  error: WebSocketEventListenerMap["error"][];
+  message: WebSocketEventListenerMap["message"][];
+  open: WebSocketEventListenerMap["open"][];
+  close: WebSocketEventListenerMap["close"][];
 };
 
 export default class ReconnectingWebSocket {
@@ -219,12 +268,12 @@ export default class ReconnectingWebSocket {
   /**
    * An event listener to be called when the WebSocket connection's readyState changes to CLOSED
    */
-  public onclose: ((event: Events.CloseEvent) => void) | null = null;
+  public onclose: ((event: CloseEvent) => void) | null = null;
 
   /**
    * An event listener to be called when an error occurs
    */
-  public onerror: ((event: Events.ErrorEvent) => void) | null = null;
+  public onerror: ((event: ErrorEvent) => void) | null = null;
 
   /**
    * An event listener to be called when a message is received from the server
@@ -292,9 +341,9 @@ export default class ReconnectingWebSocket {
   /**
    * Register an event handler of a specific event type
    */
-  public addEventListener<T extends keyof Events.WebSocketEventListenerMap>(
+  public addEventListener<T extends keyof WebSocketEventListenerMap>(
     type: T,
-    listener: Events.WebSocketEventListenerMap[T]
+    listener: WebSocketEventListenerMap[T]
   ): void {
     if (this._listeners[type]) {
       // @ts-expect-error we need to fix event/listerner types
@@ -304,7 +353,7 @@ export default class ReconnectingWebSocket {
 
   public dispatchEvent(event: Event) {
     const listeners =
-      this._listeners[event.type as keyof Events.WebSocketEventListenerMap];
+      this._listeners[event.type as keyof WebSocketEventListenerMap];
     if (listeners) {
       for (const listener of listeners) {
         this._callEventListener(event, listener);
@@ -316,9 +365,9 @@ export default class ReconnectingWebSocket {
   /**
    * Removes an event listener
    */
-  public removeEventListener<T extends keyof Events.WebSocketEventListenerMap>(
+  public removeEventListener<T extends keyof WebSocketEventListenerMap>(
     type: T,
-    listener: Events.WebSocketEventListenerMap[T]
+    listener: WebSocketEventListenerMap[T]
   ): void {
     if (this._listeners[type]) {
       // @ts-expect-error we need to fix event/listerner types
@@ -492,9 +541,9 @@ export default class ReconnectingWebSocket {
     this._retryCount = 0;
   }
 
-  private _callEventListener<T extends keyof Events.WebSocketEventListenerMap>(
-    event: Events.WebSocketEventMap[T],
-    listener: Events.WebSocketEventListenerMap[T]
+  private _callEventListener<T extends keyof WebSocketEventListenerMap>(
+    event: WebSocketEventMap[T],
+    listener: WebSocketEventListenerMap[T]
   ) {
     if ("handleEvent" in listener) {
       // @ts-expect-error we need to fix event/listerner types
@@ -539,7 +588,7 @@ export default class ReconnectingWebSocket {
     );
   };
 
-  private _handleError = (event: Events.ErrorEvent) => {
+  private _handleError = (event: ErrorEvent) => {
     this._debug("error event", event.message);
     this._disconnect(
       undefined,
@@ -557,7 +606,7 @@ export default class ReconnectingWebSocket {
     this._connect();
   };
 
-  private _handleClose = (event: Events.CloseEvent) => {
+  private _handleClose = (event: CloseEvent) => {
     this._debug("close event");
     this._clearTimeouts();
