+++ packages/partysocket/tsconfig.extract.json
@@ -2,7 +2,7 @@
 
 {
   "extends": "./tsconfig.json",
-  "include": ["./src/index.ts", "./src/react.ts"],
+  "include": ["./src/index.ts", "./src/react.ts", "./src/ws.ts"],
   "compilerOptions": {
     "declaration": true,
     "declarationMap": true,
