+++ packages/y-partykit/src/index.ts
@@ -311,6 +311,7 @@ export type YPartyKitOptions = {
    * disable gc when using snapshots!
    * */
   gc?: boolean;
+  onCommand?: (command: string, doc: WSSharedDoc) => void | Promise<void>;
   persist?: boolean;
   callback?: {
     url: string;
@@ -332,11 +333,13 @@ export function onConnect(
   const doc = getYDoc(room, options);
   doc.conns.set(conn, new Set());
   // listen and reply to events
-  conn.addEventListener("message", (message) => {
+  conn.addEventListener("message", async (message) => {
     if (typeof message.data !== "string") {
       return messageListener(conn, doc, new Uint8Array(message.data));
-    } else if (message.data === "pong") {
-      console.warn("Received non-binary message:", message.data);
+    } else if (options.onCommand) {
+      await options.onCommand(message.data, doc);
+    } else {
+      console.warn("unhandled message", message.data);
     }
   });
